# Dynamically compile shaders with uniforms instead of attributes for DDS