--- a/src/mbgl/programs/program_parameters.hpp
+++ b/src/mbgl/programs/program_parameters.hpp
@@ -7,16 +7,15 @@ namespace mbgl {
 
 class ProgramParameters {
 public:
-    ProgramParameters(float pixelRatio_ = 1.0,
-                      bool overdraw_ = false,
-                      std::string cacheDir_ = "")
-        : pixelRatio(pixelRatio_), overdraw(overdraw_), cacheDir(std::move(cacheDir_)) {
-    }
+    ProgramParameters(float pixelRatio, bool overdraw, std::string cacheDir);
 
-    const float pixelRatio;
-    const bool overdraw;
+    const std::string defines;
+
+    std::string cachePath(const char* name) const;
+
+private:
+    const std::size_t hash;
     const std::string cacheDir;
 };
 
 } // namespace mbgl
-