--- a/src/mbgl/renderer/painters/painter_fill.cpp
+++ b/src/mbgl/renderer/painters/painter_fill.cpp
@@ -38,7 +38,7 @@ void Painter::renderFill(PaintParameters& parameters,
                          const auto& drawMode,
                          const auto& indexBuffer,
                          const auto& segments) {
-            program.draw(
+            program.get(properties).draw(
                 context,
                 drawMode,
                 depthModeForSublayer(sublayer, gl::DepthMode::ReadWrite),
@@ -86,7 +86,7 @@ void Painter::renderFill(PaintParameters& parameters,
                          const auto& drawMode,
                          const auto& indexBuffer,
                          const auto& segments) {
-            program.draw(
+            program.get(properties).draw(
                 context,
                 drawMode,
                 depthModeForSublayer(sublayer, gl::DepthMode::ReadWrite),