In https://github.com/mapbox/mapbox-gl-native/commit/141e995806576364d185626176c1b993fc519291, we changed the way how most values are passed to the shader from uniforms to attributes. We did this, because for DDS, we need the ability to specify a different value for every feature, e.g. to have data-driven color values. Previously, once draw call used a uniform (constant) to give all features in a draw call the same color. In https://github.com/mapbox/mapbox-gl-native/commit/141e995806576364d185626176c1b993fc519291, we moved to constant vertex attributes, which allows us to optionally specify vertex attribute buffers for every possible combination of DDS values. Most of the time though, these attributes are set to constants.

Constant vertex attributes are an edge case features, and the [OpenGL wiki has this to say](https://www.khronos.org/opengl/wiki/Vertex_Specification#Non-array_attribute_values) about them:

> **Note:** It is not recommended that you use these. The performance characteristics of using fixed attribute data are unknown, and it is not a high-priority case that OpenGL driver developers optimize for. They might be faster than uniforms, or they might not.

Unfortunately, using constant vertex attributes causes a performance regression on some devices. To verify that this is the case, I created an isolated benchmark to find out what the culprit is. It consisted of ~36,000 vertices. The color shader received 6 colors as input and combined them to one color with `mix`. I varied the way the colors are piped through to the fragment shader in the following three combinations:

1. `uniform vec4 u_colorN` directly in the *fragment shader*: **36 fps**
2. `uniform vec4 u_colorN` in the *vertex shader*, with a `varying vec4 v_colorN` to pass the colors from the vertex to the fragment shader: **30fps**
3. `attribute vec4 a_colorN` in the vertex shader, with a `varying vec4 v_colorN` to pass the colors from the vertex to the fragment shader, with constant vertex attributes: **26 fps**

These results make it look like this performance regression is a combination of two issues (at least on the iPad mini 2):

* Passing varyings from vertex to fragment shader costs performance
* Constant vertex attributes are slower than uniforms

It's worth noting, that we had a very similar issue on WebGL in IE: https://github.com/mapbox/mapbox-gl-js/issues/1714. In JS, we're now using a different implementation that generates shader code at run-time, and uses uniforms when the DDS attribute is constant.

Therefore, I believe we'll have to switch our DDS code to a similar model that JS uses where we dynamically generate shader code at runtime based on whether the stylesheet uses data-driven values for a certain property.

I think the way we could support this is by adding `#defines` into the shader code that recompiling the shader with certain defines based on whether an attribute uses a data-driven value or not.

* [x] add `#defines` in the generated shader code to selectively toggle between attribute and uniform
* [x] add corresponding uniforms to `DataDrivenPaintProperty`-derived classes
* [x] set the uniforms as an alternative to attributes for constant data-driven values
* [x] use the correct shader with the right uniforms enabled

/cc @jfirebaugh