@jfirebaugh 2017-06-05
```
     VM SIZE                               FILE SIZE
 ++++++++++++++ GROWING                 ++++++++++++++
  +1.7% +43.5Ki __TEXT,__text           +43.5Ki  +1.7%
   +10% +10.4Ki __TEXT,__cstring        +10.4Ki   +10%
  +1.6% +2.69Ki [None]                  +3.49Ki  +2.0%
  +0.9% +2.58Ki __TEXT,__gcc_except_tab +2.58Ki  +0.9%
  +1.2%    +748 __TEXT,__unwind_info       +748  +1.2%
  +0.1%    +128 __TEXT,__const             +128  +0.1%
  +0.3%     +48 __DATA,__data               +48  +0.3%

 -------------- SHRINKING               --------------
  -0.2%     -10 __TEXT,__stub_helper        -10  -0.2%
  -0.2%      -8 __DATA,__la_symbol_ptr        0  [ = ]
  -0.2%      -6 __TEXT,__stubs               -6  -0.2%

  +1.7% +60.0Ki TOTAL                   +60.8Ki  +1.7%
```