@jfirebaugh 2017-06-05
Implementation notes:

* We still generate attribute locations for all paint attributes, even if they aren't used by the shader (this is explicitly allowed by the OpenGL spec). With more work, we could generate attribute locations for only the paint attributes that are needed. This is what GL JS does. It permits another approach to situations in which we are up against the 8 attribute limit: defer the failure case to runtime, when someone tries to use too many data-driven attributes simultaneously.
* `ConstantAttributeBinding` and related code is still present. Its use in `PaintPropertyBinders` is essentially a no-op, but it's also used in `SymbolSizeBinder`. Ideally we remove it all.