@jfirebaugh 2017-06-05
Bench results (iPhone 6) before:

```
2017-06-05 16:02:09.942756-0700 Bench GL[3994:1000717] Result:
2017-06-05 16:02:09.943195-0700 Bench GL[3994:1000717] | paris      | 52.4 fps |
2017-06-05 16:02:09.943295-0700 Bench GL[3994:1000717] | paris2     | 60.0 fps |
2017-06-05 16:02:09.943381-0700 Bench GL[3994:1000717] | alps       | 34.0 fps |
2017-06-05 16:02:09.946679-0700 Bench GL[3994:1000717] | us east    | 59.9 fps |
2017-06-05 16:02:09.946789-0700 Bench GL[3994:1000717] | greater la | 48.2 fps |
2017-06-05 16:02:09.946875-0700 Bench GL[3994:1000717] | sf         | 59.5 fps |
2017-06-05 16:02:09.946956-0700 Bench GL[3994:1000717] | oakland    | 57.8 fps |
2017-06-05 16:02:09.947037-0700 Bench GL[3994:1000717] | germany    | 30.7 fps |
2017-06-05 16:02:09.947157-0700 Bench GL[3994:1000717] Total FPS: 402.4
2017-06-05 16:02:09.947231-0700 Bench GL[3994:1000717] Average FPS: 50.3
```

After:

```
2017-06-05 15:52:51.406604-0700 Bench GL[3985:998731] Benchmark completed.
2017-06-05 15:52:51.406674-0700 Bench GL[3985:998731] Result:
2017-06-05 15:52:51.406756-0700 Bench GL[3985:998731] | paris      | 59.9 fps |
2017-06-05 15:52:51.406848-0700 Bench GL[3985:998731] | paris2     | 60.1 fps |
2017-06-05 15:52:51.406932-0700 Bench GL[3985:998731] | alps       | 40.5 fps |
2017-06-05 15:52:51.407014-0700 Bench GL[3985:998731] | us east    | 59.9 fps |
2017-06-05 15:52:51.407095-0700 Bench GL[3985:998731] | greater la | 60.1 fps |
2017-06-05 15:52:51.407177-0700 Bench GL[3985:998731] | sf         | 60.1 fps |
2017-06-05 15:52:51.407257-0700 Bench GL[3985:998731] | oakland    | 60.0 fps |
2017-06-05 15:52:51.407338-0700 Bench GL[3985:998731] | germany    | 39.3 fps |
2017-06-05 15:52:51.407410-0700 Bench GL[3985:998731] Total FPS: 439.8
2017-06-05 15:52:51.407481-0700 Bench GL[3985:998731] Average FPS: 55.0
```