@kkaefer 2017-06-06
Here are my benchmarks on the iPad mini 2:

| benchmark  | before DDS | after DDS | master | dynamic shaders
| --- | --- | --- | --- | ---
| paris      | 30.7 fps | 16.4 fps | 17.8 fps | 22.2 fps
| paris2     | 32.3 fps | 20.5 fps | 19.6 fps | 27.3 fps
| alps       | 16.2 fps |  7.8 fps |  8.8 fps | 14.0 fps
| us east    | 40.0 fps | 27.3 fps | 25.6 fps | 33.2 fps
| greater la | 29.5 fps | 17.1 fps | 17.9 fps | 24.9 fps
| sf         | 43.2 fps | 25.3 fps | 23.6 fps | 33.5 fps
| oakland    | 31.0 fps | 19.2 fps | 19.9 fps | 25.3 fps
| germany    | 18.6 fps |  9.0 fps | 10.1 fps | 16.2 fps

Looking at the shaders, it looks like there are still attributes generated for some of the shaders even though they're `#ifdef`ed out. Maybe calling `glBindAttribLocation` has an adverse effect?