@kkaefer 2017-06-06
Compare this with the pre-DDS state of the same program:

![bench gl gputrace 2017-06-06 16-10-43](https://cloud.githubusercontent.com/assets/52399/26833465/d07d4098-4ad2-11e7-9425-61e86e9f5f77.png)