@jfirebaugh 2017-06-06
@kkaefer Can you run bench on iPad mini 2 one more time? I rebased on master, including 9dfc2d924d440560adb2db13c758b2c5b3b7dd47, which is a significant per-frame performance speedup.