@kkaefer 2017-06-06
Results are pretty much unchanged:

| benchmark | performance |
| --- | --- |
| paris      | 21.0 fps |
| paris2     | 27.7 fps |
| alps       | 12.6 fps |
| us east    | 30.0 fps |
| greater la | 20.2 fps |
| sf         | 33.7 fps |
| oakland    | 23.1 fps |
| germany    | 14.5 fps |