@kkaefer 2017-06-13
Remaining task is to convert the size attributes for symbol programs to this format as well.