@jfirebaugh 2017-06-13
Capturing from chat:

Porting symbol size to dynamic shaders is difficult. It doesn't fit in the existing `#pragma` system. That's why it wasn't implemented that way to begin with. I think we should merge and port the current changes, and not try to switch this right now.

@ansis:

> that sounds good to me. The current plan for working around the 8 attribute limit for pitch-scaling is to pack a_size into a_data
> if a_size was changed to a dynamically generated separate attribute (which I guess it should be eventually) we'd need to find another way

I profiled the changes here and found #9257. With that included, FPS is maxed at 60 on my iPhone.